requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-Limiter
Flask-Caching
Flask-Mail
python-dotenv
email-validator
gunicorn
psycopg2-binary
